+++ aitune/torch/jit/tune.py
@@ -1,0 +1,16 @@
+# SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+"""Module for tuning JIT models."""
+
+from aitune.torch.jit.config import JITMode, config
+from aitune.torch.jit.patcher import Patcher
+from aitune.torch.utils.memory import cleanup_memory
+
+
+def deferred():
+    """Tune the modules in deferred mode."""
+    if config.mode != JITMode.TUNE_DEFERRED:
+        raise ValueError(f"tune.deferred() requires JITMode.TUNE_DEFERRED, but current mode is {config.mode.value}")
+
+    cleanup_memory()
+    Patcher.tune_deferred()