+++ tests/functional/pytorch/jit/006_jit_sd15_inspect.py
@@ -35,8 +35,8 @@ def test_jit_sd15():
 
     def batch():
         with torch.no_grad():
-            pipe([prompt] * 1, num_inference_steps=1)
-            pipe([prompt] * 2, num_inference_steps=1)
+            pipe([prompt] * 1)
+            pipe([prompt] * 2)
 
     for _ in range(5):
         batch()